+++ foundry/scripts/core/constants.js
@@ -1,7 +1,9 @@
 export const MODULE_ID = "bane-of-azeroth";
 
+export const ADVENTURE_PACK_NAME = "bane-of-azeroth"; // BOA_REBRAND_PRESERVE
+
 export const ADVENTURE_PACK_ID =
-  `${MODULE_ID}.${MODULE_ID}`;
+  `${MODULE_ID}.${ADVENTURE_PACK_NAME}`;
 
 export const ADVENTURE_PROMPT_VERSION_SETTING =
   "adventurePromptVersion";